refactor: separé tipo articulo en 3
- articulo (base)
- articuloInsumo
- articuloManifacturado
