+++ src/services/types/articulo.ts
@@ -16,48 +16,3 @@ export interface Articulo extends BaseEntity {
   unidadDeMedida: UnidadDeMedidaRef
   categoria: Categoria
 }
-
-/** Artículo Insumo */
-export interface ArticuloInsumo extends Articulo {
-  precioCompra: number
-  stockActual: number
-  stockMaximo: number
-  stockMinimo: number
-  esParaElaborar: boolean
-}
-
-/** Artículo Manufacturado — se completará en la siguiente etapa */
-export interface ArticuloManufacturado extends Articulo {
-  descripcion: string
-  tiempoEstimadoMinutos: number
-  preparacion: string
-  detalles: Array<{cantidad: number; idArticuloInsumo: number}>
-}
-
-/** Inputs — tal como espera la API */
-export interface ArticuloInsumoCreateInput {
-  denominacion: string
-  precioVenta: number
-  precioCompra: number
-  stockActual: number
-  stockMaximo: number
-  stockMinimo: number
-  esParaElaborar: boolean
-  idSucursal: number
-  idUnidadDeMedida: number
-  idCategoria: number
-}
-
-export interface ArticuloInsumoUpdateInput {
-  denominacion?: string
-  precioVenta?: number
-  precioCompra?: number
-  stockActual?: number
-  stockMaximo?: number
-  stockMinimo?: number
-  esParaElaborar?: boolean
-  // Relaciones (confirmar con backend si son editables)
-  idUnidadDeMedida?: number
-  idCategoria?: number
-  idSucursal?: number // probablemente inmutable
-}
